TOTPQrCodeGenerationException was moved to security.tfa package
